% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe.1cfull}
\alias{delta.optim.scqe.1cfull}
\title{Delta optimization method for \code{scqe} 1 cohort, full data}
\usage{
delta.optim.scqe.1cfull(treatment, outcome, delta, obj, specified = NULL,
  ...)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{print} method provides the critical values presented in the summary method for scqe objects.
}
