% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe2}
\alias{delta.optim.scqe2}
\title{Delta optimization method for \code{scqe} 2 cohort, full data}
\usage{
delta.optim.scqe2(post, treatment, outcome, delta, obj, alpha,
  specified = NULL, ...)
}
\arguments{
\item{post}{Indicator for being in the second cohort.}

\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1.}

\item{outcome}{Continueous vector representing the outcome for each
observation}

\item{delta}{Can take either a single value or vector of possible values.}

\item{obj}{scqe object.}

\item{alpha}{Numeric alpha for confidence intervals (default alpha=.05).}

\item{specified}{specified additional arguments.}

\item{...}{Extra optional arguments}
}
\description{
The \code{print} method provides the critical values presented in the summary
method for scqe objects.
}
