% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.scqe}
\alias{plot.scqe}
\title{Plot method for \code{scqe}}
\usage{
\method{plot}{scqe}(scqe.obj, ...)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{print} method provides...
}
\examples{
# give example here

}
