% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe-package}
\alias{scqe-package}
\title{scqe: Stability Controlled Quasi-Experimentation}
\description{
The scqe package contains several function for statistical analysis that
factor in confounding variables and their impact on estimates
(Hazlett, 2019).
}
\details{
The main function in the package is \code{\link{scqe}}, which computes scqe
estimates and confidence intervals for one or two cohorts with summary or
full data given.
}
\section{package dependencies}{

AER
ggplot2
}

\references{
Hazlett, C. (2019), 'Estimating causal effects of new treatments
despite self-selection: The case of experimental medical treatments.'
Journal of Causal Inference.
}
