% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe.1cfull}
\alias{scqe.1cfull}
\title{Stability controlled quasi-experiment (scqe) for 1 cohort case with FULL DATA}
\usage{
scqe.1cfull(treatment, outcome, delta, ...)
}
\arguments{
\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}
}
\description{
This function returns the scqe estimates, standard deviations,  confidence
intervals for the one cohort case. This function is similar to the main scqe function
but does not have 2 cohort (ie there is only one measurement time (no "post" input)).
}
\examples{
# Put examples here!
set.seed(1234)
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .25, size=1)


scqe_1cohort.out = scqe_1cohort(treatment=tx, outcome=y, delta=c(-0.1,0,.1))
plot(scqe_1cohort.out)
summary(scqe_1cohort.out)



}
