% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe_master}
\alias{scqe_master}
\title{Stability controlled quasi-experiment (scqe)}
\usage{
scqe_master(post, treatment, outcome, delta, cohort, untr_pre, untr_post,
  tr_post, tr_pre, Y_tr_post, Y_untr_post, Y_tr_pre, Y_untr_pre, untr, tr,
  Y_tr, Y_untr, min_delta, max_delta)
}
\arguments{
\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}

\item{cohort}{1 or 2 depending existence of time element.}

\item{untr_pre}{Integer number of untreated patients in the first cohort if applicable (T=0).}

\item{untr_post}{Integer number of untreated patients in the second cohort if applicable (T=1).}

\item{tr_post}{Integer number of treated patients in the second cohort if applicable (T=1).}

\item{Y_tr_post}{Outcome for patients who received treatment at time T=1.}

\item{Y_untr_post}{Outcome for patients who did not receive treatment at time T=1.}

\item{Y_tr_pre}{Outcome for patients who did receive treatment at time T=0.}

\item{Y_untr_pre}{Outcome for patients who did not receive treatment at time T=0.}

\item{untr}{Integer number of untreated patients.}

\item{tr}{Integer number of treated patients.}

\item{Y_tr}{Outcome for treated patients.}

\item{min_delta}{Minimum delta.}

\item{max_delta}{Maximum delta.}

\item{tre_pre}{Integer number of treated patients in the first cohort if applicable (T=0).}
}
\description{

}
\examples{
set.seed(1234)
post = c(rep(0,100), rep(1,100))
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .1 + .02*post - .05*tx, size=1)

2 COHORT, FULL DATA
scqe_master.out = scqe_master(post=post, treatment=tx, outcome=y, delta=c(-0.1,0,.1), cohort=2)
plot(scqe_master.out)
summary(scqe_master.out)

1 COHORT, FULL DATA
scqe_master.out = scqe_master(treatment=tx, outcome=y, delta=c(-0.1,0,.1), cohort=1)
plot(scqe_master.out)
summary(scqe_master.out)

1 COHORT, SUMMARY STATS
scqe_master.out = scqe_master(untr=100,tr=200,Y_untr=5,Y_tr=50,min_delta=.1,max_delta=1,cohort=1)
plot(scqe_master.out)
summary(scqe_master.out)

2 COHORT, SUMMARY STATS
scqe_master.out = scqe_master(untr_pre=200,untr_post=150,tr_post=50,tr_pre=0,Y_tr_post=20,Y_untr_post=1,Y_tr_pre=0,Y_untr_pre=5,min_delta=.1, max_delta=1,cohort=2)
plot(scqe_master.out)
summary(scqe_master.out)

ALL OF THESE MATCH THE SHINY APP RESULTS

}
\references{
Hazlett, C. (2019), "Estimating causal effects of new treatments despite self-selection: The case of experimental medical treatments. Journal of Causal Inference.
}
