% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqemethod}
\alias{scqemethod}
\title{Stability controlled quasi-experiment (scqe)}
\usage{
scqemethod(...)
}
\arguments{
\item{...}{Arguments from scqe}
}
\value{
scqe object of class "scqe", results table
}
\description{
Dispatches to correct scqe function
}
