% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.scqe}
\alias{summary.scqe}
\title{Summary method for \code{scqe}}
\usage{
\method{summary}{scqe}(scqe.obj)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{summary} method provides...
}
\examples{
set.seed(1234)
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .25, size=1)

scqe_1cohort.out = scqe_1cohort(treatment=tx, outcome=y, delta=c(-0.1,0,.1))

plot(scqe_1cohort.out)
summary(scqe_1cohort.out)

}
