% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe.1cfull}
\alias{delta.optim.scqe.1cfull}
\title{Delta optimization method for \code{scqe} 1 cohort, full data}
\usage{
delta.optim.scqe.1cfull(treatment, outcome, delta, obj, specified = NULL,
  alpha, ...)
}
\arguments{
\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values.}

\item{obj}{scqe object}

\item{specified}{specified}

\item{alpha}{Numeric alpha for confidence intervals (default alpha=.05).}

\item{...}{Extra optional arguments}
}
\description{
The \code{print} method provides the critical values presented in the summary
method for scqe objects.
}
