% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe}
\alias{delta.optim.scqe}
\title{Delta optimization method for \code{scqe}}
\usage{
delta.optim.scqe(Y_T0, untreated, Y_untreated, treated, Y_treated, obj,
  specified = NULL, ...)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{print} method provides...
}
