% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe}
\alias{delta.optim.scqe}
\title{Delta optimization method for \code{scqe} 1 cohort, summary statistics}
\usage{
delta.optim.scqe(
  Y_T0,
  untreated,
  Y_untreated,
  treated,
  Y_treated,
  obj,
  specified = NULL,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{Y_T0}{Y}

\item{untreated}{Number of untreated individuals.}

\item{Y_untreated}{Outcome for untreated individuals.}

\item{treated}{Number of treated individuals.}

\item{Y_treated}{Outcome for treated individuals.}

\item{obj}{scqe object.}

\item{specified}{Specified optional arguments.}

\item{alpha}{Numeric alpha for confidence intervals (default alpha = 0.05).}

\item{...}{Extra optional arguments.}
}
\value{
Optimal delta.
}
\description{
The \code{print} method provides the critical values presented in the summary
method for scqe objects.
}
