% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe2}
\alias{delta.optim.scqe2}
\title{Delta optimization method for \code{scqe} 2 cohort, full data}
\usage{
delta.optim.scqe2(post, treatment, outcome, delta, obj, alpha,
  specified = NULL, ...)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{print} method provides the critical values presented in the summary
method for scqe objects.
}
