% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta.optim.scqe2}
\alias{delta.optim.scqe2}
\title{Delta optimization method for \code{scqe} 2 cohort, full data}
\usage{
delta.optim.scqe2(
  post,
  treatment,
  outcome,
  delta,
  obj,
  alpha = 0.05,
  specified = NULL,
  ...
)
}
\arguments{
\item{post}{Binary vector corresponding to T=0,1 for each observation.}

\item{treatment}{Binary or continuous vector corresponding (usually) to 0,1.}

\item{outcome}{Continuous vector representing the outcome for each
observation}

\item{delta}{Single value or vector of possible values for change in
average non-treatment outcome between cohorts (if applicable).}

\item{obj}{scqe object.}

\item{alpha}{Numeric alpha for confidence intervals (default alpha=.05).}

\item{specified}{Specified optional arguments.}

\item{...}{Extra optional arguments.}
}
\value{
Optimal delta.
}
\description{
The \code{print} method provides the critical values presented in the summary
method for scqe objects.
}
