% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta_optim_SCQE_2C}
\alias{delta_optim_SCQE_2C}
\title{Delta optimization method for \code{scqe} 2 cohort, summary statistics}
\usage{
delta_optim_SCQE_2C(delta, untr_pre, untr_post, tr_post, tr_pre, Y_tr_post,
  Y_untr_post, Y_tr_pre, Y_untr_pre, obj, specified = NULL, ...)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{print} method provides the critical values presented in the summary method for scqe objects.
}
