% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe.1cfull}
\alias{scqe.1cfull}
\title{Stability controlled quasi-experiment (scqe) for 1 cohort case, full data}
\usage{
scqe.1cfull(treatment, outcome, delta, alpha, ...)
}
\arguments{
\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}

\item{alpha}{Numeric alpha for confidence interval (default is alpha=.05).}
}
\description{
This function returns the scqe estimates and confidence intervals for the
1 cohort case (ie there is not 'post' input) when the user inputs full data.
}
\examples{
set.seed(1234)
post = c(rep(0,100), rep(1,100))
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .1 + .02*post - .05*tx, size=1)

# One cohort, full data
scqe.1cohort.full = scqe(treatment=tx, outcome=y,
delta=seq(from=-.1, to=.1, by=0.05))
plot(scqe.1cohort.full)
summary(scqe.1cohort.full)



}
