% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe.1csumm}
\alias{scqe.1csumm}
\title{Stability controlled quasi-experiment (scqe) for 1 cohort case SUMMARY STATISTICS AS INPUT}
\usage{
\method{scqe}{1csumm}(untr_1C, Y_untr_1C, tr_1C, Y_tr_1C, min_outcome,
  max_outcome, ...)
}
\arguments{
\item{untr_1C}{Number of unreated individuals.}

\item{Y_untr_1C}{Outcome for untreated individuals.}

\item{tr_1C}{Number of treated individuals.}

\item{Y_tr_1C}{Outcome for treated individuals.}

\item{min_outcome}{Minimum possible delta.}

\item{max_outcome}{Maximum possible delta.}
}
\description{
This function returns the scqe estimates, standard deviations,  confidence
intervals for the one cohort case. This function is similar to the main scqe function
but does not have 2 cohorts (ie there is only one measurement time (no "post" input)).
It takes in summary statistics about the data instead of the full data.
}
\examples{
# Put examples here!
set.seed(1234)
one_cohort_scqe.out <- one_cohort_scqe(200, 5, 50, 25, .1, 1)
plot(one_cohort_scqe.out)
summary(one_cohort_scqe.out)


}
\references{
Hazlett, C. (?)
}
