% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe.2cfull}
\alias{scqe.2cfull}
\title{Stability controlled quasi-experiment (scqe) for 2 cohort case, full data}
\usage{
scqe.2cfull(post, treatment, outcome, delta, ...)
}
\arguments{
\item{post}{Binary vector corresponding to T=0,1 for each observation.}

\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}
}
\description{
This function returns the scqe estimates and confidence intervals for the
2 cohort case when the user inputs full data.
}
\examples{
set.seed(1234)
post = c(rep(0,100), rep(1,100))
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .1 + .02*post - .05*tx, size=1)

# Two cohorts, full data
scqe.2cohort.full = scqe(scqe.obj, post=post, treatment=tx, outcome=y, delta=seq(from=-.1,to=.1, by=0.05))
scqe.2cohort.full$result
plot(scqe.2cohort.full)
summary(scqe.2cohort.full)

}
