% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/scqe.R
\docType{package}
\name{scqe}
\alias{scqe}
\alias{scqe-package}
\title{SCQE estimates.}
\usage{
scqe(post, treatment, outcome, delta)
}
\arguments{
\item{post}{Binary vector corresponding to T=0,1 for each observation.}

\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}
}
\description{
\code{scqe} returns the scqe estimates, standard deviations, and relevant confidence
intervals.

The scqe package contains several function for statistical analysis that factor in
confounding variables and their impact on estimates (Hazlett, 2019).
}
\section{scqe functions}{

scqeE: gets the scqe estimates, sd, and CIs
}

\section{package dependencies}{

AER
}

