% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe}
\alias{scqe}
\title{Stability controlled quasi-experiment (scqe)}
\usage{
scqe(post = NA, treatment = NA, outcome = NA, delta = NA,
  cohort = NA, untr_pre = NA, untr_post = NA, tr_post = NA,
  tr_pre = NA, Y_tr_post = NA, Y_untr_post = NA, Y_tr_pre = NA,
  Y_untr_pre = NA, untr = NA, tr = NA, Y_tr = NA, Y_untr = NA,
  min_delta = NA, max_delta = NA, alpha = NA, ...)
}
\arguments{
\item{post}{Indicator for being in the second cohort}

\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}

\item{cohort}{1 or 2 depending existence of time element.}

\item{untr_pre}{Integer number of untreated patients in the first cohort if applicable (T=0).}

\item{untr_post}{Integer number of untreated patients in the second cohort if applicable (T=1).}

\item{tr_post}{Integer number of treated patients in the second cohort if applicable (T=1).}

\item{Y_tr_post}{Outcome for patients who received treatment at time T=1.}

\item{Y_untr_post}{Outcome for patients who did not receive treatment at time T=1.}

\item{Y_tr_pre}{Outcome for patients who did receive treatment at time T=0.}

\item{Y_untr_pre}{Outcome for patients who did not receive treatment at time T=0.}

\item{untr}{Integer number of untreated patients.}

\item{tr}{Integer number of treated patients.}

\item{Y_tr}{Outcome for treated patients.}

\item{min_delta}{Minimum delta.}

\item{max_delta}{Maximum delta.}

\item{alpha}{Numeric alpha for confidence interval.}

\item{tre_pre}{Integer number of treated patients in the first cohort if applicable (T=0).}
}
\description{
Main scqe function. Computes scqe estimates and corresponding confidence intervals.
}
\examples{
set.seed(1234)
post = c(rep(0,100), rep(1,100))
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .1 + .02*post - .05*tx, size=1)

# Two cohorts, full data
scqe.2cohort.full = scqe(post=post, treatment=tx, outcome=y, delta=seq(from=-.1, to=.1, by=0.05))
scqe.2cohort.full$result
plot(scqe.2cohort.full)
summary(scqe.2cohort.full)

# One cohort, full data
scqe.1cohort.full = scqe(treatment=tx, outcome=y, delta=seq(from=-.1, to=.1, by=0.05))
scqe.1cohort.full$result
plot(scqe.1cohort.full)
summary(scqe.1cohort.full)

# Two cohorts, summary data only
scqe.2cohort.sum = scqe(untr_pre=200,untr_post=150,tr_post=50,tr_pre=0,Y_tr_post=20,Y_untr_post=1,Y_tr_pre=0, Y_untr_pre=5,min_delta=.1, max_delta=1)
scqe.2cohort.sum$result
plot(scqe.2cohort.sum)
summary(scqe.2cohort.sum)

# One cohort, summary data only
scqe.1cohort.sum = scqe(untr=100,tr=200,Y_untr=5,Y_tr=50,min_outcome=.1,max_outcome=1, min_delta=.1,max_delta=1)
scqe.1cohort.sum$result
plot(scqe.1cohort.sum)
summary(scqe.1cohort.sum)

}
\references{
Hazlett, C. (2019), "Estimating causal effects of new treatments despite self-selection: The case of experimental medical treatments." Journal of Causal Inference.
}
