% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scqe.R
\docType{package}
\name{scqe}
\alias{scqe}
\alias{scqe-package}
\title{scqe: Stability Controlled Quasi-Experimentation}
\description{
The scqe package contains several function for statistical analysis that factor in
confounding variables and their impact on estimates (Hazlett, 2019).
}
\section{scqe functions}{

getSCQE: gets the scqe estimates, sd, and CIs
}

\section{package dependencies}{

AER
}

