% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scqe}
\alias{scqe}
\title{Stability controlled quasi-experiment (scqe) for 2 cohort case FULL DATA}
\usage{
scqe(post, treatment, outcome, delta, ...)
}
\arguments{
\item{post}{Binary vector corresponding to T=0,1 for each observation.}

\item{treatment}{Binary or continuous vector correspoding (usually) to 0,1
(no treatment or treatment) for each observation.}

\item{outcome}{Continuous vector representing the outcome for each
observation.}

\item{delta}{Can take either a single value or vector of possible values for
delta.}
}
\description{
This function returns the scqe estimates, standard deviations,  confidence
intervals,...
}
\examples{
# Put examples here!
set.seed(1234)
post = c(rep(0,100), rep(1,100))
tx = c(rep(0, 100), rbinom(n = 100, prob=.27, size=1))
y = rbinom(n=200, prob = .1 + .02*post - .05*tx, size=1)

scqe.out = scqe(post=post, treatment=tx, outcome=y, delta=c(-0.1,0,.1))

plot(scqe.out)
summary(scqe.out)

}
\references{
Hazlett, C. (2019), "Estimating causal effects of new treatments despite self-selection: The case of experimental medical treatments. Journal of Causal Inference.
}
