% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.scqe}
\alias{summary.scqe}
\title{Summary method for \code{scqe}}
\usage{
\method{summary}{scqe}(scqe.obj)
}
\arguments{
\item{scqe.obj}{an object of class \code{\link{scqe}}}
}
\description{
The \code{summary} method provides...
}
\examples{
# give example here

}
